server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["electric lies <electric.lies@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
black = "^23.12.1"
pycryptodome = "^3.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
